+++ test/lightning/credentials/audit_test.exs
@@ -15,8 +15,9 @@ defmodule Lightning.Credentials.AuditTest do
         Audit.event("created", credential.id, user.id)
         |> Audit.save()
 
-      assert audit.row_id == credential.id
-      assert %{before: nil, after: nil} = audit.metadata
+      assert audit.item_type == "credential"
+      assert audit.item_id == credential.id
+      assert %{before: nil, after: nil} = audit.changes
       assert audit.event == "created"
       assert audit.actor_id == user.id
     end
@@ -33,17 +34,40 @@ defmodule Lightning.Credentials.AuditTest do
         Audit.event("updated", credential.id, user.id, changeset)
         |> Audit.save()
 
-      assert audit.row_id == credential.id
+      assert audit.item_type == "credential"
+      assert audit.item_id == credential.id
 
       # Check that the body attribute is encrypted in audit records too.
       # We can only test the beginning of the encrypted string as the
       # algorithm include randomness or padding of some sort.
       # %{}
-      assert audit.metadata.before.body =~ "AQpBRVMuR0NNLlYx"
+      assert audit.changes.before.body =~ "AQpBRVMuR0NNLlYx"
       # %{"my-secret" => "value}
-      assert audit.metadata.after.body =~ "AQpBRVMuR0NNLlYx"
+      assert audit.changes.after.body =~ "AQpBRVMuR0NNLlYx"
 
       assert audit.event == "updated"
+      assert audit.actor_id == user.id
+    end
+
+    test "generates 'deleted' audit trail entries" do
+      user = user_fixture()
+
+      credential =
+        credential_fixture(user_id: user.id, body: %{"my-secret" => "value"})
+
+      {:ok, audit} =
+        Audit.event("deleted", credential.id, user.id)
+        |> Audit.save()
+
+      assert audit.item_type == "credential"
+      assert audit.item_id == credential.id
+
+      assert audit.changes == %Lightning.Credentials.Audit.Changes{
+               before: nil,
+               after: nil
+             }
+
+      assert audit.event == "deleted"
       assert audit.actor_id == user.id
     end
   end